+++ android/app/src/main/java/app/priceguard/ui/util/Dialog.kt
@@ -3,7 +3,6 @@ package app.priceguard.ui.util
 import android.os.Bundle
 import androidx.appcompat.app.AppCompatActivity
 import androidx.fragment.app.Fragment
-import app.priceguard.ui.ErrorDialogFragment
 import app.priceguard.ui.data.DialogConfirmAction
 
 fun AppCompatActivity.showConfirmDialog(
