+++ android/app/src/main/java/app/priceguard/ui/util/ErrorDialogFragment.kt
@@ -1,4 +1,4 @@
-package app.priceguard.ui
+package app.priceguard.ui.util
 
 import android.app.Dialog
 import android.content.Intent
